add texture-based simplex noise terrain rendering
some experimentation with procedural mesh generation and texturing. the
terrain is split into individual chunks and each chunk is rendered as
one mesh. an integral based vertex system is used which should, in
theory, save time & space, but we will see if that change sticks around

much of this code (in particular, the mesh generation code) is
suboptimal, but it will do for now until a permanent and efficient
solution is found