add gpu_context abstraction
this type serves as the primary interface to the gpu. additional work
needs to be done to further decouple render_context and gpu_context, but
it is better than it was. most critically, gpu memory management is
still too manual.

this is the first step toward runtime swappable shaders. the next step
is likely a render pipeline builder abstraction to allow easy
customization of render pipelines at runtime